pyproject.toml
[project]
name = "qwen3-vl"
readme = "ReadME.md"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
    "datasets>=4.0.0",
    "dotenv>=0.9.9",
    "einops>=0.8.1",
    "huggingface-hub>=0.34.2",
    "numpy<3",
    "pillow>=11.3.0",
    "torch==2.8.0",
    "torchvision>=0.22.1",
    "transformers>=4.54.0",
    "wandb>=0.21.0",
    "rich>=13.3.4",
    "matplotlib>=3.10.7",
    "accelerate>=1.12.0",
    "pytest>=9.0.2",
    "imageio[ffmpeg]>=2.37.2",
    "gradio>=6.3.0",
]


[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py312"
fix = true
